requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pandas>=1.5.0
plotly>=5.0.0
python-dotenv>=1.0.0
Pillow>=9.0.0
supabase>=2.0.0
streamlit-authenticator>=0.2.0
requests>=2.28.0
